requirements.txt
pandas
numpy
xgboost
scikit-learn
matplotlib
seaborn
python-dotenv
openai
langchain
langchain-community
faiss-cpu
streamlit
huggingface-hub
tiktoken
